requirements.txt
Django == 1.8
oauthlib == 1.0.1
requests == 2.7.0
python-social-auth == 0.2.12
django-braces == 1.8.1
django-taggit == 0.16.2
django-autocomplete-light == 2.2.5
